requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
openai>=1.10.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.25.0
pytest-cov>=4.1.0
httpx>=0.26.0
python-multipart>=0.0.9
